pyproject.toml
[project]
name = "f1-predictor"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "django-cors-headers==4.7.0",
    "djangorestframework==3.16.0",
    "python-dotenv==1.1.1",
    "requests==2.32.4",
]
